FastFeastDAGs/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-common-sql
apache-airflow-providers-standard
apache-airflow-providers-snowflake+
python-dotenv